+++ backend/models/problem.go
@@ -1,6 +1,12 @@
 package models
 
+import (
+	"backend/db"
+	"database/sql"
+)
+
 type StatusType string
+type SolveStatusType string
 type DifficultyType string
 
 type Example struct {
@@ -12,9 +18,9 @@ type Example struct {
 }
 
 const (
-	Unsolved   StatusType = "unsolved"
-	Processing StatusType = "processing"
-	Solved     StatusType = "solved"
+	Unsolved   SolveStatusType = "unsolved"
+	Processing SolveStatusType = "processing"
+	Solved     SolveStatusType = "solved"
 )
 
 const (
@@ -41,7 +47,7 @@ type Problem struct {
 type ProblemListItem struct {
 	Id         int            `json:"id"`
 	Name       string         `json:"name"`
-	Status     StatusType     `json:"status"`
+	Status     SolveStatusType     `json:"status"`
 	Answers    int            `json:"answers"`
 	PassRate   int            `json:"pass_rate"`
 	Difficulty DifficultyType `json:"difficulty"`
@@ -53,4 +59,34 @@ type ProblemItem struct {
 	Description string         `json:"description"`
 	Difficulty  DifficultyType `json:"difficulty"`
 	Examples    []Example      `json:"examples"`
+}
+
+type UserProblemListItem struct {
+	Id         int            `json:"id"`
+	Name       string         `json:"name"`
+	Status     StatusType     `json:"status"`
+	Difficulty DifficultyType `json:"difficulty"`
+}
+
+func AddProblem(name string, description string, difficulty DifficultyType, createdBy int) error {
+	var _ int64
+	var lastInsertId int64
+
+	err := db.DB.QueryRow("INSERT INTO problems(name, description, difficulty, created_by) VALUES($1, $2, $3, $4) RETURNING id", name, description, difficulty, createdBy).Scan(&lastInsertId)
+
+	_ = lastInsertId
+
+	return err
+}
+
+func GetProblems() (*sql.Rows, error) {
+	rows, err := db.DB.Query("SELECT id, name, difficulty FROM problems")
+
+	return rows, err
+}
+
+func GetProblemsByUserId(userId int) (*sql.Rows, error) {
+	rows, err := db.DB.Query("SELECT id, name, difficulty, status FROM problems WHERE created_by = $1", userId)
+
+	return rows, err
 }