+++ frontend/src/components/Table.tsx
@@ -1,7 +1,6 @@
 import { Table as DefaultTable } from "@radix-ui/themes";
 import { ReactNode } from "react";
 import { styled } from "styled-components";
-import Loading from "./Loading";
 
 interface IColumn {
   key: string;
@@ -11,19 +10,7 @@ interface IColumn {
 interface ITable {
   columns: IColumn[];
   children?: ReactNode;
-  loading?: boolean;
 }
-
-const Wrapper = styled.div`
-  position: fixed;
-  top: 0;
-  left: 0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 500px;
-`;
 
 export const TableCell = styled(DefaultTable.Cell)<{ maxwidth?: number }>`
   overflow: hidden;
@@ -32,8 +19,9 @@ export const TableCell = styled(DefaultTable.Cell)<{ maxwidth?: number }>`
   max-width: ${(props) => props.maxwidth}px;
 `;
 
-
-export const TableRowHeaderCell = styled(DefaultTable.RowHeaderCell)<{ maxwidth?: number }>`
+export const TableRowHeaderCell = styled(DefaultTable.RowHeaderCell)<{
+  maxwidth?: number;
+}>`
   overflow: hidden;
   text-overflow: ellipsis;
   white-space: nowrap;
@@ -55,11 +43,6 @@ export default function Table(props: ITable) {
         </DefaultTable.Header>
         <DefaultTable.Body>{props.children}</DefaultTable.Body>
       </DefaultTable.Root>
-      {props.loading && (
-        <Wrapper>
-          <Loading visible={props.loading}>加载中</Loading>
-        </Wrapper>
-      )}
     </>
   );
 }