+++ frontend/src/constants.ts
@@ -4,7 +4,7 @@ export const AppName = import.meta.env.VITE_APP_NAME;
 export const AppNameEn = import.meta.env.VITE_APP_NAME_EN;
 export const smallBreakpoints = ["initial", "xs", "sm"];
 export const difficultyMap: IBadge["map"] = {
-  easy: { label: "简单", color: "crimson" },
+  easy: { label: "简单", color: "green" },
   medium: { label: "中等", color: "orange" },
-  hard: { label: "困难", color: "green" },
+  hard: { label: "困难", color: "crimson" },
 };